improved test output
